This is my certificate of online course of Excel.
This course is the best online course that I have ever attended. All the concepts were clearly explained by the teacher. Moreover, the quality of the videos was also perfect. Everyone who wants to learn about basics of excel should do this course. Since it is free of cost, you could leave it any time if you do not like it.